+++ src/rectobject.cpp
@@ -599,8 +599,8 @@ static PyObject* rect_richcompare(PyObject* a, PyObject* b, int op) {
 static PyObject* rect_repr(PyObject* self) {
   Rect* x = ((RectObject*)self)->m_x;
   return PyString_FromFormat("<gameracore.Rect offset_y: %i offset_x: %i nrows: %i ncols: %i>",
-			     x->offset_y(), x->offset_x(), x->nrows(),
-			     x->ncols());
+			     (int)x->offset_y(), (int)x->offset_x(),
+			     (int)x->nrows(), (int)x->ncols());
 }
 
 void init_RectType(PyObject* module_dict) {
